int -> intc to avoid platform issues

+++ pande_gas/utils/rdkit_utils.py
@@ -128,7 +128,7 @@ def generate_conformers(mol, n_conformers=1, rmsd_threshold=0.5):
     cids = AllChem.EmbedMultipleConfs(mol, n_conformers,
                                       pruneRmsThresh=rmsd_threshold)
     assert mol.GetNumConformers() >= 1
-    cids = np.asarray(cids, dtype=int)
+    cids = np.asarray(cids, dtype='intc')
 
     # minimize conformers and get energies
     energy = np.zeros(cids.size, dtype=float)
